+++ weatherly/enums.py
@@ -175,3 +175,9 @@ class SportsEventType(Enum):
     football = 1
     cricket = 2
     golf = 3
+
+    def __str__(self):
+        return f"<SportsEventType: {self.name}>"
+
+    def __repr__(self):
+        return repr(self.__str__())