+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 📜 Weatherly changelog
-In this file you can view the changelog, including updates and changes that were made to the package.
+In this file you can view the changelog, including updates and changes that were made to this package.
+
+### Version 0.7.0
+This version adds supports for the Sports API. Following things were added:
+* `weatherly.Client(...).get_sports_data` for fetching sports data
+* `SportsData` as a response from function above and `SportsEvent` that is a single event, part of the `SportsData` class. 
 
 ### Version 0.6.0
 This version adds support for the Astronomy API and IP Lookup API. Two new methods for the `weatherly.Client` were implemented: